// types/product.ts
export interface ProductAttributes {
    name: string
    storage: number
    image: string 
    display: string 
    camera: string
    battery: string
    ram: number
    processor: string
    warranty: number 
    price: number
    assured: boolean 
    discount: number
    rating: number
    brand: string 
    popularity: number
  }